requirements.txt
absl-py==1.4.0
alembic==1.11.1
astunparse==1.6.3
CacheControl==0.13.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
firebase-admin==6.1.0
Flask==2.2.3
Flask-Admin==1.6.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==4.0.4
flask-serialize==2.1.3
Flask-SQLAlchemy==3.0.3
flatbuffers==23.5.26
gast==0.4.0
google-api-core==2.11.1
google-api-python-client==2.89.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.54.2
h5py==3.8.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
jax==0.4.10
Jinja2==3.1.2
keras==2.12.0
libclang==16.0.0
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
msgpack==1.0.5
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
permissive-dict==1.0.4
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.3
pyparsing==3.0.9
python-dotenv==1.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.13
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.14.1
WTForms==3.0.1
